fix: use u32::MAX for network graph reset to avoid LDK early return
LDK's remove_stale_channels_and_tracking_with_time() returns immediately
if current_time > u32::MAX. Passing u64::MAX meant the graph was never
actually cleared. Use u32::MAX instead so all channels appear stale.

Also includes Swift bindings for listClosedChannels and ClosedChannelDetails.

+++ src/lib.rs
@@ -1784,8 +1784,10 @@ impl Node {
 			})?;
 		}
 
-		// Clear the in-memory graph by marking all channels as stale
-		self.network_graph.remove_stale_channels_and_tracking_with_time(u64::MAX);
+		// Clear the in-memory graph by marking all channels as stale.
+		// Note: LDK's implementation returns early if current_time > u32::MAX,
+		// so we pass u32::MAX to ensure all channels appear stale and get removed.
+		self.network_graph.remove_stale_channels_and_tracking_with_time(u32::MAX as u64);
 
 		Ok(())
 	}
